refactor(compensation): refactor GlobalDrawer component and update related components
- Update import path for GlobalDrawer and useGlobalDrawer
- Replace 'content' prop with 'children' in Drawer components
- Refactor GlobalDrawer component to use DrawerProps directly
- Create separate file for useGlobalDrawer hook
- Update components that use GlobalDrawer to match new API
