+++ compensation/dashboard/src/features/Failed/ApplyRetrySpec.tsx
@@ -14,7 +14,7 @@
 import { Button, Form, InputNumber, App, Input } from "antd";
 import type { RetrySpec } from "../../services";
 import { executionFailedCommandService } from "../../services/executionFailedCommandClient.ts";
-import { useGlobalDrawer } from "../../components/GlobalDrawer/GlobalDrawer.tsx";
+import { useGlobalDrawer } from "../../components/GlobalDrawer";
 import type { OnChangedCapable } from "./Actions.tsx";
 
 export interface ApplyRetrySpecProps extends OnChangedCapable {
